requirements.txt
feedparser==6.0.12
Flask==3.1.3
newspaper==0.1.0.7
python-dotenv==1.2.2
SQLAlchemy==2.0.48